refactor(lanes): extract arch dialect (loc/cz_partner/location attrs) from movement
Move the architecture-level location primitives out of the gemini movement
dialect into a new lanes-level dialect at `bloqade.lanes.dialects.arch`
(dialect name "bloqade.lanes.arch"): the `Loc`, `CzPartner`, `WordId`,
`SiteId`, `ZoneId` statements plus their interface (`loc` / `cz_partner`),
const-prop (`CzPartnerConstProp`), rewrites (`RewriteLocationAttr`,
`BindCzPartnerArchSpec`), and interpreter impls.

These depend only on `LocationAddress` + `ArchSpec` (no gemini deps), so they
belong at the lanes layer. The movement dialect keeps only the user-movement
statements (`MoveTo`, `Permute`) and imports `LocationAddressType` from arch.
The physical kernel group now unions both movement and arch dialects.

Also fixes two pre-existing lanes -> gemini upward imports: `upstream.py` and
`pipeline/base.py` now import `BindCzPartnerArchSpec` from
`bloqade.lanes.dialects.arch` instead of `bloqade.gemini...movement.rewrite`.
