+++ python/bloqade/gemini/common/dialects/movement/__init__.py
@@ -1,8 +1,3 @@
-from . import constprop as constprop, rewrite as rewrite, stmts as stmts
+from . import stmts as stmts
 from ._dialect import dialect as dialect
-from ._interface import (
-    cz_partner as cz_partner,
-    loc as loc,
-    move_to as move_to,
-    permute as permute,
-)
+from ._interface import move_to as move_to, permute as permute
