+++ python/bloqade/gemini/common/impl/__init__.py
@@ -1,6 +1,5 @@
 from . import (
     duplicate_address as duplicate_address,
-    movement as movement,
     qubit_address as qubit_address,
     terminal_measure as terminal_measure,
 )
