+++ python/bloqade/gemini/logical/__init__.py
@@ -1,4 +1,5 @@
-from bloqade.gemini.common.dialects.movement import loc as loc, move_to as move_to
+from bloqade.gemini.common.dialects.movement import move_to as move_to
+from bloqade.lanes.dialects.arch import loc as loc
 
 from . import dialects as dialects, impl as impl, validation as validation
 from .dialects.operations import terminal_measure as terminal_measure
