+++ python/bloqade/gemini/physical/group.py
@@ -13,14 +13,17 @@ from typing_extensions import Doc
 from bloqade.gemini import common as gemini_common
 from bloqade.gemini.logical.group import kernel as logical_kernel
 from bloqade.lanes.arch.spec import ArchSpec
+from bloqade.lanes.dialects import arch as arch_dialect
 
 
-@ir.dialect_group(logical_kernel.union([gemini_common.dialects.movement]))
+@ir.dialect_group(logical_kernel.union([gemini_common.dialects.movement, arch_dialect]))
 def kernel(self):
     """Gemini kernel with user-directed atom movement.
 
-    Extends the logical ``kernel`` with the movement dialect; use this when
-    your kernel calls ``movement.move_to(...)``.
+    Extends the logical ``kernel`` with the movement dialect (``move_to`` /
+    ``permute``) and the ``bloqade.lanes.arch`` dialect (``loc`` / ``cz_partner``
+    / location-attribute reads); use this when your kernel calls
+    ``movement.move_to(...)`` / location primitives.
     """
     address_analysis = address.AddressAnalysis(dialects=self)
