+++ python/bloqade/lanes/dialects/arch/_dialect.py
@@ -1,0 +1,3 @@
+from kirin import ir
+
+dialect = ir.Dialect(name="bloqade.lanes.arch")
