+++ python/bloqade/lanes/upstream.py
@@ -12,11 +12,11 @@ from kirin.ir.method import Method
 from kirin.rewrite.abc import RewriteRule
 
 from bloqade.gemini.common.dialects import qubit as gemini_qubit
-from bloqade.gemini.common.dialects.movement.rewrite import BindCzPartnerArchSpec
 from bloqade.gemini.logical.rewrite.initialize import _RewriteU3ToInitialize
 from bloqade.lanes.analysis import layout, placement
 from bloqade.lanes.arch.spec import ArchSpec
 from bloqade.lanes.dialects import move, place
+from bloqade.lanes.dialects.arch import BindCzPartnerArchSpec
 from bloqade.lanes.passes import SequentialPlacePass
 from bloqade.lanes.rewrite import circuit2place, place2move, resolve_pinned, state
